crate.version: Move `downloadData()` property to `DownloadGraph` component
