+++ app/templates/crate/version.hbs
@@ -100,7 +100,7 @@
     </div>
     <div local-class='graph'>
       <h4>Downloads over the last 90 days</h4>
-      <DownloadGraph @data={{this.downloadData}} local-class="graph-data" />
+      <DownloadGraph @data={{this.downloads}} local-class="graph-data" />
     </div>
   </div>
 {{/unless}}
